Alow a new Service Provider option to disallow user login.
In real life use, not every user in Symfony Application can login in
your Service Providers.

This patch makes available a new option for service providers ,
'userAllowed', which can be defined as a callable and makes it posible
to define which users will be able to login to the SP.

When a user is not allowed, an UserNotAllowedInServiceProvider
exception will be thrown. It is a 403 symfony exception, but
includes the service provider variable, which can be use in a
custom error403.html.twig.

Also, a new config has been added to Service Providers called
'description' to get a human readable name for each SP.